Moved interfaces from types.ts to store.ts so they are closer to their use
